requirements.txt
streamlit
openai
google-generativeai
psycopg2-binary
pandas
numpy
python-dotenv
xlsxwriter
pillow
